rename proposal adapter, move helper fns into utils dir
